refactor(page): extract useArrangeManager hook (phase 3)
Move the arrange/duplicate subsystem — settings state, blocking-overlay
state + timer, high-precision/auto/array arrange orchestration, the
~515-line packing algorithm, and the duplicate apply/fill handlers — out
of page.tsx into features/scene/arrange/useArrangeManager.ts. State lives
in the hook and is consumed in Home via destructure-same-names, so
ArrangePanel/DuplicatePanel/hotkeys/scene-preview consumers are unchanged.

The shared support-aware geometry helpers (getArrangeTransform,
getModelSupportAwareDimensionsMm, footprint-polygon ref,
supportBoundsByModelId) stay in Home — they're also used by the
gizmo/transform system — and are injected as deps. arrangeSpacingMm stays
in Home too (the editor context menu reads it) and is injected.

page.tsx: 21,604 -> 20,195 lines (-1,409). Verified via `tsc --noEmit`.
Worker-based packing — smoke-test arrange/duplicate before merge.
